+++ upload/library/Iversia/FAQ/Search/DataHandler/Question.php
@@ -1,0 +1,5 @@
+<?php
+
+class Iversia_FAQ_Search_DataHandler_Question extends XenForo_Search_DataHandler_Abstract
+{
+}